refactor(factory): read the messages-pending state from one source
The favicon fix drilled `messagesInitializing` into ChatTranscriptProvider,
which duplicated a fact ChatThreadMessagesContext already held. The context and
its hook now live in their own modules — matching the convention every other
chat context follows — and the provider sits above the transcript, so the
stepper, the composer and the favicon all read the same pending flag.

Also drops `| null` from the favicon prop in favour of an optional one, and
gives ThreadPage's resolving branch a name instead of an inline fragment.
