+++ mastracode/factory-ui/src/ui/domains/chat/context/ChatThreadMessagesContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+export interface ChatThreadMessagesApi {
+  threadId?: string;
+  isPending: boolean;
+  error: unknown;
+}
+
+export const ChatThreadMessagesContext = createContext<ChatThreadMessagesApi | null>(null);
